Some code refactoring

+++ hardware/1Wire.cpp
@@ -21,7 +21,7 @@
 
 #define round(a) ( int ) ( a + .5 )
 
-C1Wire::C1Wire(const int ID, const int sensorThreadPeriod, const int switchThreadPeriod, const std::string &path)
+C1Wire::C1Wire(const int ID, const int sensorThreadPeriod, const int switchThreadPeriod, const std::string& path)
 	: m_system(nullptr)
 	, m_sensorThreadPeriod(sensorThreadPeriod)
 	, m_switchThreadPeriod(switchThreadPeriod)
@@ -142,13 +142,13 @@ void C1Wire::SensorThread()
 	if (m_sensors.empty())
 		return; // quit if no sensors
 
-	pollPeriod = m_sensorThreadPeriod / m_sensors.size();
+	pollPeriod = (int)(m_sensorThreadPeriod / m_sensors.size());
 
 	if (pollPeriod > 1000)
 	{
 		pollIterations = pollPeriod / 1000;
 		pollPeriod = 1000;
-		afterIterations = (m_sensorThreadPeriod / 1000) - (pollIterations*m_sensors.size());
+		afterIterations = (int)((m_sensorThreadPeriod / 1000) - (pollIterations * m_sensors.size()));
 	}
 
 	// initial small delay
@@ -162,7 +162,7 @@ void C1Wire::SensorThread()
 		}
 
 		// Parse our devices
-		for (const auto &device : m_sensors)
+		for (const auto& device : m_sensors)
 		{
 			if (IsStopRequested(0))
 				break;
@@ -277,9 +277,9 @@ void C1Wire::SwitchThread()
 }
 
 
-bool C1Wire::WriteToHardware(const char *pdata, const unsigned char /*length*/)
+bool C1Wire::WriteToHardware(const char* pdata, const unsigned char /*length*/)
 {
-	const tRBUF *pSen = reinterpret_cast<const tRBUF*>(pdata);
+	const tRBUF* pSen = reinterpret_cast<const tRBUF*>(pdata);
 
 	if (!m_system)
 		return false;//no 1-wire support
@@ -304,16 +304,16 @@ void C1Wire::BuildSensorList() {
 		return;
 
 	std::vector<_t1WireDevice> devices;
-	int n1wireVersion=0;
+	int n1wireVersion = 0;
 
 	_log.Debug(DEBUG_HARDWARE, "1-Wire: Searching sensors");
 
 	m_sensors.clear();
 	m_system->GetDevices(devices);
 
-	if (!m_sql.GetPreferencesVar("1-Wire-version",n1wireVersion)) n1wireVersion=0;
+	if (!m_sql.GetPreferencesVar("1-Wire-version", n1wireVersion)) n1wireVersion = 0;
 
-	for (const auto &device : devices)
+	for (const auto& device : devices)
 	{
 		switch (device.family)
 		{
@@ -332,7 +332,7 @@ void C1Wire::BuildSensorList() {
 
 				DeviceIdToByteArray(device.devid, deviceIdByteArray);
 				uint16_t oldID = (deviceIdByteArray[0] << 8) | deviceIdByteArray[1];
-				uint16_t newID = ((deviceIdByteArray[0]^deviceIdByteArray[2]^deviceIdByteArray[4]) << 8) | (deviceIdByteArray[1]^deviceIdByteArray[3]^deviceIdByteArray[5]);
+				uint16_t newID = ((deviceIdByteArray[0] ^ deviceIdByteArray[2] ^ deviceIdByteArray[4]) << 8) | (deviceIdByteArray[1] ^ deviceIdByteArray[3] ^ deviceIdByteArray[5]);
 
 				m_sql.safe_query("UPDATE DeviceStatus SET DeviceID='%d', Unit='%d' WHERE (HardwareID=='%d') AND (DeviceID=='%d') AND (Unit=='%d') AND (Type=='%d') AND (SubType=='%d')",
 					newID, newID & 0xff, m_HwdID, oldID, oldID & 0xff, pTypeTEMP, sTypeTEMP5);
@@ -352,7 +352,7 @@ void C1Wire::BuildSensorList() {
 	}
 	devices.clear();
 
-	m_sql.UpdatePreferencesVar("1-Wire-version",1);
+	m_sql.UpdatePreferencesVar("1-Wire-version", 1);
 
 }
 
@@ -367,7 +367,7 @@ void C1Wire::BuildSwitchList() {
 	m_switches.clear();
 	m_system->GetDevices(devices);
 
-	for (const auto &device : devices)
+	for (const auto& device : devices)
 	{
 		switch (device.family)
 		{
@@ -395,7 +395,7 @@ void C1Wire::PollSwitches()
 		return;
 
 	// Parse our devices (have to test if m_TaskSwitches.IsStopRequested because it can take some time in case of big networks)
-	for (const auto &device : m_switches)
+	for (const auto& device : m_switches)
 	{
 		// Manage families specificities
 		switch (device.family)
@@ -478,7 +478,7 @@ void C1Wire::ReportTemperature(const std::string& deviceId, const float temperat
 
 	//XOR all bytes to ID instead of last two
 	//uint16_t lID = (deviceIdByteArray[0] << 8) | deviceIdByteArray[1];
-	uint16_t lID = ((deviceIdByteArray[0]^deviceIdByteArray[2]^deviceIdByteArray[4]) << 8) | (deviceIdByteArray[1]^deviceIdByteArray[3]^deviceIdByteArray[5]);
+	uint16_t lID = ((deviceIdByteArray[0] ^ deviceIdByteArray[2] ^ deviceIdByteArray[4]) << 8) | (deviceIdByteArray[1] ^ deviceIdByteArray[3] ^ deviceIdByteArray[5]);
 
 	SendTempSensor(lID, 255, temperature, "Temperature");
 }
@@ -551,7 +551,7 @@ void C1Wire::ReportVoltage(const std::string& /*deviceId*/, const int unit, cons
 
 	tsen.RFXSENSOR.msg1 = (BYTE)(voltage / 256);
 	tsen.RFXSENSOR.msg2 = (BYTE)(voltage - (tsen.RFXSENSOR.msg1 * 256));
-	sDecodeRXMessage(this, (const unsigned char *)&tsen.RFXSENSOR, nullptr, 255, nullptr);
+	sDecodeRXMessage(this, (const unsigned char*)&tsen.RFXSENSOR, nullptr, 255, nullptr);
 }
 
 void C1Wire::ReportIlluminance(const std::string& deviceId, const float illuminescence)
